Centralize DB settings in misago.conf
