+++ misago/users/admin/forms.py
@@ -683,175 +683,3 @@ class FilterDataDownloadsForm(forms.Form):
             )
 
         return queryset
-
-
-class ChangeUsersSettingsForm(ChangeSettingsForm):
-    settings = [
-        "account_activation",
-        "allow_custom_avatars",
-        "avatar_upload_limit",
-        "default_avatar",
-        "default_gravatar_fallback",
-        "signature_length_max",
-        "subscribe_reply",
-        "subscribe_start",
-        "username_length_max",
-        "username_length_min",
-    ]
-
-    account_activation = forms.ChoiceField(
-        label=_("Require new accounts activation"),
-        choices=[
-            ("none", _("No activation required")),
-            ("user", _("Activation token sent to user e-mail")),
-            ("admin", _("Activation by administrator")),
-            ("closed", _("Disable new registrations")),
-        ],
-        widget=forms.RadioSelect(),
-    )
-    username_length_min = forms.IntegerField(
-        label=_("Minimum allowed username length"), min_value=2, max_value=20
-    )
-    username_length_max = forms.IntegerField(
-        label=_("Maximum allowed username length"), min_value=2, max_value=20
-    )
-    allow_custom_avatars = YesNoSwitch(
-        label=_("Allow custom avatar uploads"),
-        help_text=_(
-            "Turning this option off will forbid forum users from uploading custom "
-            "avatars. Good for forums adressed at young users."
-        ),
-    )
-    avatar_upload_limit = forms.IntegerField(
-        label=_("Maximum size of uploaded avatar"),
-        help_text=_("Enter maximum allowed file size (in KB) for avatar uploads."),
-        min_value=0,
-    )
-    default_avatar = forms.ChoiceField(
-        label=_("Default avatar"),
-        choices=[
-            ("dynamic", _("Individual")),
-            ("gravatar", _("Gravatar")),
-            ("gallery", _("Random avatar from gallery")),
-        ],
-        widget=forms.RadioSelect(),
-    )
-    default_gravatar_fallback = forms.ChoiceField(
-        label=_("Fallback for default gravatar"),
-        help_text=_(
-            "Select which avatar to use when user has no gravatar associated with "
-            "their e-mail address."
-        ),
-        choices=[
-            ("dynamic", _("Individual")),
-            ("gallery", _("Random avatar from gallery")),
-        ],
-        widget=forms.RadioSelect(),
-    )
-    signature_length_max = forms.IntegerField(
-        label=_("Maximum allowed signature length"), min_value=10, max_value=5000
-    )
-    subscribe_start = forms.ChoiceField(
-        label=_("Started threads"),
-        choices=[
-            ("no", _("Don't watch")),
-            ("watch", _("Put on watched threads list")),
-            (
-                "watch_email",
-                _("Put on watched threads list and e-mail user when somebody replies"),
-            ),
-        ],
-        widget=forms.RadioSelect(),
-    )
-    subscribe_reply = forms.ChoiceField(
-        label=_("Replied threads"),
-        choices=[
-            ("no", _("Don't watch")),
-            ("watch", _("Put on watched threads list")),
-            (
-                "watch_email",
-                _("Put on watched threads list and e-mail user when somebody replies"),
-            ),
-        ],
-        widget=forms.RadioSelect(),
-    )
-
-
-class ChangeCaptchaSettingsForm(ChangeSettingsForm):
-    settings = [
-        "captcha_type",
-        "recaptcha_site_key",
-        "recaptcha_secret_key",
-        "qa_question",
-        "qa_help_text",
-        "qa_answers",
-    ]
-
-    captcha_type = forms.ChoiceField(
-        label=_("Enable CAPTCHA"),
-        choices=[
-            ("no", _("No CAPTCHA")),
-            ("re", _("reCaptcha")),
-            ("qa", _("Question and answer")),
-        ],
-        widget=forms.RadioSelect(),
-    )
-    recaptcha_site_key = forms.CharField(
-        label=_("Site key"), max_length=100, required=False
-    )
-    recaptcha_secret_key = forms.CharField(
-        label=_("Secret key"), max_length=100, required=False
-    )
-    qa_question = forms.CharField(
-        label=_("Test question"), max_length=100, required=False
-    )
-    qa_help_text = forms.CharField(
-        label=_("Question help text"), max_length=250, required=False
-    )
-    qa_answers = forms.CharField(
-        label=_("Valid answers"),
-        help_text=_("Enter each answer in new line. Answers are case-insensitive."),
-        widget=forms.Textarea({"rows": 4}),
-        max_length=250,
-        required=False,
-    )
-
-    def clean(self):
-        cleaned_data = super().clean()
-
-        if cleaned_data.get("captcha_type") == "re":
-            if not cleaned_data.get("recaptcha_site_key"):
-                self.add_error(
-                    "recaptcha_site_key",
-                    _(
-                        "You need to enter site key if "
-                        "selected CAPTCHA type is reCaptcha."
-                    ),
-                )
-
-            if not cleaned_data.get("recaptcha_secret_key"):
-                self.add_error(
-                    "recaptcha_secret_key",
-                    _(
-                        "You need to enter secret key if "
-                        "selected CAPTCHA type is reCaptcha."
-                    ),
-                )
-
-        if cleaned_data.get("captcha_type") == "qa":
-            if not cleaned_data.get("qa_question"):
-                self.add_error(
-                    "qa_question",
-                    _("You need to set question if selected CAPTCHA type is Q&A."),
-                )
-
-            if not cleaned_data.get("qa_answers"):
-                self.add_error(
-                    "qa_answers",
-                    _(
-                        "You need to set question answers if "
-                        "selected CAPTCHA type is Q&A."
-                    ),
-                )
-
-        return cleaned_data
